refactor: move linux kernel api compat to kernel_compat.h
